# Conflicts:
#	src/main/java/com/projects/cardpayment/controllers/CardController.java

+++ src/main/java/com/projects/cardpayment/service/CardService.java
@@ -11,6 +11,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 
 import com.projects.cardpayment.constant.CardPaymentConstants;
@@ -30,6 +31,7 @@ import com.itextpdf.text.Paragraph;
 import com.itextpdf.text.pdf.PdfWriter;
 import com.projects.cardpayment.entities.Card;
 import com.projects.cardpayment.entities.TxnDetails;
+import com.projects.cardpayment.entities.User;
 import com.projects.cardpayment.repository.CardRepository;
 import com.projects.cardpayment.repository.TxnRepository;
 import com.projects.cardpayment.utils.mailservice.MailService;
@@ -49,24 +51,23 @@ public class CardService {
 	@Autowired
 	private MailService mailService;
 
-	public Card createCard(Card card) {
+	@Autowired
+	private UserAppService userAppService;
 
+	public Card createCard(Card card) {
 		logger.info("CPA : CS : Saving card details {}", card);
 		Card savedCard = null;
-
 		try {
 			savedCard = cardRepository.save(card);
 		} catch (Exception e) {
 			logger.error("Exception found at CPA : CS ", e);
 		}
 		return savedCard;
-
 	}
 
 	public List<Card> getAllCards() {
 		logger.info("inside getallCards service method");
 		List<Card> allCards = cardRepository.findAll();
-
 		logger.info("CS : number of cards fetched from db {}", allCards);
 		return allCards;
 	}
@@ -77,8 +78,56 @@ public class CardService {
 		return card;
 	}
 
-	public void deleteById(Integer cardId) {
-		cardRepository.deleteById(cardId);
+	public Map<String, String> deleteById(Integer cardId, String userName, String password) {
+		logger.info("VALIDATING USERNAME AND PASSWORD userName:{},psssword:{}", userName, password);
+		final Map<String, String> deleteCardAPIResponse = new HashMap<>();
+		User user = userAppService.userLoginService(userName, password);
+		if (user != null) {
+			logger.info("Login validation successful");
+			String role = user.getRole();
+			if (role.equalsIgnoreCase("ADMIN")) {
+				Optional<Card> cardid = cardRepository.findById(cardId);
+				if (cardid.isPresent()) {
+					logger.info("Admin validation successfull");
+					cardRepository.deleteById(cardId);
+
+					deleteCardAPIResponse.put(CardPaymentConstants.STATUS, CardPaymentConstants.SUCCESS);
+					deleteCardAPIResponse.put(CardPaymentConstants.STATUS_CODE,
+							CardPaymentConstants.SUCCESS_STATUS_200);
+					deleteCardAPIResponse.put(CardPaymentConstants.STATUS_MSG, "Card deleted successfully");
+					deleteCardAPIResponse.put("cardId", String.valueOf(cardId));
+					logger.info("1 card deleted");
+					return deleteCardAPIResponse;
+				} else {
+					logger.info("card not found");
+					deleteCardAPIResponse.put(CardPaymentConstants.STATUS, CardPaymentConstants.FAILURE);
+					deleteCardAPIResponse.put(CardPaymentConstants.STATUS_CODE,
+							CardPaymentConstants.FAILURE_STATUS_500);
+					deleteCardAPIResponse.put(CardPaymentConstants.STATUS_MSG, "Card not found");
+					deleteCardAPIResponse.put("cardId", String.valueOf(cardId));
+					deleteCardAPIResponse.put(CardPaymentConstants.REASON, "Invalid card Id!!");
+					logger.info("card not found");
+					return deleteCardAPIResponse;
+				}
+			} else {
+				logger.info("Admin validation failed");
+				deleteCardAPIResponse.put(CardPaymentConstants.STATUS, CardPaymentConstants.FAILURE);
+				deleteCardAPIResponse.put(CardPaymentConstants.STATUS_CODE, "401");
+				deleteCardAPIResponse.put(CardPaymentConstants.STATUS_MSG, "Authorization failed");
+				deleteCardAPIResponse.put("cardId", String.valueOf(cardId));
+				deleteCardAPIResponse.put(CardPaymentConstants.REASON, "user doesn't have admin role");
+				return deleteCardAPIResponse;
+			}
+		} else {
+			logger.info("username or password incorrect");
+			deleteCardAPIResponse.put(CardPaymentConstants.STATUS, CardPaymentConstants.FAILURE);
+			deleteCardAPIResponse.put(CardPaymentConstants.STATUS_CODE, CardPaymentConstants.FAILURE_STATUS_500);
+			deleteCardAPIResponse.put(CardPaymentConstants.STATUS_MSG, "Authentication failed!!");
+			deleteCardAPIResponse.put("cardId", String.valueOf(cardId));
+			deleteCardAPIResponse.put(CardPaymentConstants.REASON, "username or password is incorrect");
+			logger.info("user doesn't exist");
+			return deleteCardAPIResponse;
+		}
 	}
 
 	public String addMoneyToCard(Integer cardId, int amount) {
@@ -178,12 +227,9 @@ public class CardService {
 
 					MailStructure mailStructure = new MailStructure();
 					mailStructure.setSubject("Order Payment from Card");
-					mailStructure.setMessage(
-							"Dear "+card.getCardHolderFirstName()+",\n"
-							+  amountToBePaid +" has been deducted(order payment) from your card successfully.\n"
-									+ "Txn ID : " + txnUUID +"\n\n"
-									+ "Thanks & Regards, \n"
-									+ "CardPaymentApp");
+					mailStructure.setMessage("Dear " + card.getCardHolderFirstName() + ",\n" + amountToBePaid
+							+ " has been deducted(order payment) from your card successfully.\n" + "Txn ID : " + txnUUID
+							+ "\n\n" + "Thanks & Regards, \n" + "CardPaymentApp");
 
 					mailService.sendEmail(mailStructure, card.getEmail());
 					logger.info("Mail has been sent to {} successfully.", card.getEmail());
@@ -221,12 +267,9 @@ public class CardService {
 
 	public Map<String, String> moneyTransfer(Integer senderCardId, Integer receiverCardId, Integer amount)
 			throws FileNotFoundException, DocumentException {
-
 		logger.info("Input received in money tranfer service");
 		logger.info("senderCardId :: {}, receiverCardId :: {}, amount :: {}", senderCardId, receiverCardId, amount);
-
 		Map<String, String> moneyTransferResponse = null;
-
 		try {
 			Integer serviceChargeAmount = 0;
 			Integer txnAmount = amount;
@@ -241,7 +284,6 @@ public class CardService {
 					Integer extraAmount = amount - 5000;
 					serviceChargeAmount = (extraAmount * 5) / 100;
 					logger.info("calculated service charged {}", serviceChargeAmount);
-
 					amount = amount - serviceChargeAmount;
 					logger.info("Final amount after service charge deduction :: {} ", amount);
 				}
@@ -294,12 +336,10 @@ public class CardService {
 			moneyTransferResponse.put("reason", "Minimum txn amount must be 100");
 		}
 		return moneyTransferResponse;
-
 	}
 
 	private String generateTxnReceiptPdf(Card senderCard, Card receiverCard, Integer txnAmount)
 			throws FileNotFoundException, DocumentException {
-
 		Document document = new Document(PageSize.LETTER);
 		String uuid = String.valueOf(UUID.randomUUID());
 		String pdfName = uuid + "_receipt.pdf";
@@ -332,9 +372,7 @@ public class CardService {
 
 	private void saveTransactionalDetails(Integer senderId, String senderName, Integer receiverId, String receiverName,
 			Integer amount, Date txnDate, String purpose, String txnUUID) {
-
 		TxnDetails txnDetails = new TxnDetails();
-
 		txnDetails.setSenderName(senderName);
 		txnDetails.setSenderId(senderId);
 		txnDetails.setReceiverName(receiverName);
@@ -343,35 +381,26 @@ public class CardService {
 		txnDetails.setTxnDate(txnDate);
 		txnDetails.setPurpose(purpose);
 		txnDetails.setTxnUUID(txnUUID);
-
 		logger.info("saveTransactionalDetails {}", txnDetails);
 		logger.info("setting txnUUID {}", txnUUID);
 		txnRepository.save(txnDetails);
 		logger.info("Transactional details save successfully");
-
 	}
 
 	public List<Card> getListOfExpiredCards() {
 		List<Card> listOfExpiredCard = new ArrayList<>();
-
 		List<Card> cards = cardRepository.findAll();
 		for (int i = 0; i < cards.size(); i++) {
 			Card card = cards.get(i);
-
 			String cardExpiryDate = card.getCardExpiryDate();
-
 			DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
-
 			// Parse the string into a LocalDate object
 			LocalDate cardDate = LocalDate.parse(cardExpiryDate, formatter);
-
 			// Compare with another date
 			LocalDate currentDate = LocalDate.now();
-
 			if (cardDate.isBefore(currentDate)) {
 				logger.info("Card is expired {}", card);
 				listOfExpiredCard.add(card);
-
 			}
 		}
 		return listOfExpiredCard;
@@ -381,9 +410,7 @@ public class CardService {
 		Integer cardCVVNo = 0;
 		Card card = cardRepository.findById(cardId).get();
 		cardCVVNo = card.getCardCVVNumber();
-
 		return cardCVVNo;
-
 	}
 
 	public List<Card> getListOfCardThatHaveBalanceInBetween(Integer lowerAmount, Integer upperAmount) {
@@ -391,21 +418,16 @@ public class CardService {
 		List<Card> listOfCardBalance = new ArrayList<>();
 		Card card = null;
 		allCards = cardRepository.findAll();
-
 		logger.info("allCards {}", allCards);
-
 		for (int i = 0; i < allCards.size(); i++) {
 			card = allCards.get(i);
 			Integer cardBalance = card.getCardBalance();
 			logger.info("card Balance*** {}", cardBalance);
-
 			if (cardBalance >= lowerAmount && cardBalance <= upperAmount) {
-
 				listOfCardBalance.add(card);
 			}
 		}
 		return listOfCardBalance;
-
 	}
 
 	public List<Card> getCardsByBankName(String bankName) {
@@ -413,7 +435,6 @@ public class CardService {
 		List<Card> listOfCards = null;
 		Card card = null;
 		String cardBankName = null;
-
 		allCards = cardRepository.findAll();
 		listOfCards = new ArrayList<>();
 		for (int i = 0; i < allCards.size(); i++) {
@@ -425,5 +446,4 @@ public class CardService {
 		}
 		return listOfCards;
 	}
-
 }
